Update authentication to be compatible with PouchDB Auth Library
Authentication mechanism is now compatible with PouchDB Auth library. Additionally, users are stored in a database. The API key password for the user is encrypted with the user's password on registration, and decrypted on login. The API key and decrypted API key password are used to authentication with Cloudant and the resultant authentication cookie is given to the client.